Added the subnet modifications, but still need to do some testing before PR.
